homework-7/app/requirements.txt
django
psycopg2-binary
python-dotenv
Pillow
djangorestframework
drf-yasg
pytest
django-pytest
pytest-django